pyproject.toml
[tool.poetry]
name = "Platform Registry API"
version = "0.0.1"
description = "API for managing platforms on the Health Data Hub"
authors = ["APIs TEAM <apis-team@aphp.fr>"]

[build-system]
requires = ["poetry==1.8.3"]
build-backend = "poetry.masonry.api"

[tool.poetry.dependencies]
python = "^3.11"
Werkzeug = "^3.0.3"
alembic = "^1.13.2"
fastapi = "^0.109.2"
httpx = "^0.27.0"
jose = "^1.0.0"
kubernetes = "^24.2.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
psycopg = "^3.2.1"
pydantic = {extras = ["email"], version = "^2.6.1"}
pydantic-settings = "^2.3.4"
pyjwt = "^2.8.0"
python-multipart = "^0.0.9"
pytest = "^8.1.1"
python-dateutil = "^2.8.2"
python-dotenv = "^1.0.1"
pytz = "^2022.1"
requests = "^2.32.3"
sqlalchemy = "^2.0.27"
starlette = "^0.36.3"
uvicorn = "^0.27.1"

[tool.poetry.group.dev.dependencies]
coverage = "^7.6.0"
pre-commit = "^3.7.1"
pyjwt = "^2.8.0"
pytest = "^8.3.1"
ruff = "^0.5.4"

[tool.ruff]
target-version = "^py311"
line-length = 150
exclude = [
    ".git",
    ".git-rewrite",
    ".ruff_cache",
    ".venv",
    "__pypackages__",
    "build",
    "venv",
    "tests",
    "alembic"
]

[tool.ruff.lint]
select = ["E", "F"]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
fixable = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE",
           "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT","ICN", "INP", "ISC", "NPY", "PD", "PGH",
           "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"]
unfixable = []
